% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{filter_list}
\alias{filter_list}
\title{Filter list elements by name}
\usage{
filter_list(l, var)
}
\arguments{
\item{l}{A named list to filter.}

\item{var}{A character vector with the names of the elements to keep.}
}
\value{
A list containing only the elements with names that match the values in the 'var' argument.
}
\description{
This function filters a named list by the elements with names matching a given character vector.
}
\examples{
my_list <- list(a = 1, b = 2, c = 3)
filter_list(my_list, c("a", "c"))

}
