% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{list_to_string}
\alias{list_to_string}
\title{Convert a list of values to a string}
\usage{
list_to_string(x)
}
\arguments{
\item{x}{A list to be converted to a string.}
}
\value{
A character vector containing the values of the list as a string.
}
\description{
This function takes a list of values and converts it to a string by applying the 'toString' function to each element of the list.
}
\examples{
list_to_string(list(1:3, "foo", TRUE))
#[1] "1, 2, 3" "foo" "TRUE"

}
