% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{load_file}
\alias{load_file}
\title{Download and load an XML file from a given URL.}
\usage{
load_file(link)
}
\arguments{
\item{link}{a character string specifying the URL from which to download the XML file}
}
\value{
the XML data read into an R session
}
\description{
This function downloads an XML file from a given URL, reads it into an R session using the xml2 package, and returns it as an R object.
}
