% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyebsco.R
\name{tidy_ebsco_xml}
\alias{tidy_ebsco_xml}
\title{Tidy EBSCOhost XML}
\usage{
tidy_ebsco_xml(file)
}
\arguments{
\item{file}{character string specifying the file path to the XML file.}
}
\value{
A tidy data frame with extracted author, affiliation, and publication information.
}
\description{
Takes an XML file exported from the literature search database EBSCOhost and extracts relevant data into a tidy data frame.
}
\examples{
# Load example data
xml_file <- system.file("extdata", "example.xml", package = "tidyebsco")

# Extract data
tidy_ebsco_xml(xml_file)

}
